Refactor code to subsample adata with a specified seed
The `evaluate` method in `CellEmbeddingsEval` now takes an additional `seed` parameter to allow for reproducible subsampling of the `adata` object. If the `adata` object has more cells than the specified `n_cells`, a subset of `n_cells` cells will be taken using the provided `seed`. This change improves the reproducibility of the evaluation process.

Also, convert sparse matrix to dense array in `InputData`

In the `InputData` class, the `adata.X` attribute is now converted to a dense array if it is a sparse matrix. This ensures that the data is in a consistent format for downstream processing.

Fix typo in `GeneExprPredEval`

A comment in the `GeneExprPredEval` class had a typo, where the word "teally" was used instead of "really". This typo has been corrected for clarity.